tools/pyproject.toml
[tool.poetry]
name = "tools"
version = "0.1.0"
description = ""
authors = ["Koutaro Mukai <mukai.k1011k@outlook.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pyserial = "^3.5"
opencv-python-headless = "^4.8.1.78"
qingpi-python = {path = "../qingpi-python"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
